requirements.txt
crewai>=0.30.0
crewai-tools>=1.14.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
